+++ corporate/views/billing_page.py
@@ -5,41 +5,15 @@ from django.http import HttpRequest, HttpResponse, HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse
 
-from corporate.lib.stripe import RealmBillingSession, UpdatePlanRequest
-from corporate.models import CustomerPlan, get_current_plan_by_customer, get_customer_by_realm
+from corporate.lib.stripe import RealmBillingSession, UpdatePlanRequest, is_realm_on_paid_plan
+from corporate.models import CustomerPlan, get_customer_by_realm
 from zerver.decorator import require_billing_access, zulip_login_required
 from zerver.lib.request import REQ, has_request_variables
 from zerver.lib.response import json_success
 from zerver.lib.validator import check_int, check_int_in, check_string
-from zerver.models import Realm, UserProfile
+from zerver.models import UserProfile
 
 billing_logger = logging.getLogger("corporate.stripe")
-
-PAID_PLANS = [
-    Realm.PLAN_TYPE_STANDARD,
-    Realm.PLAN_TYPE_PLUS,
-]
-
-
-def is_realm_on_paid_plan(realm: Realm) -> bool:
-    return realm.plan_type in PAID_PLANS
-
-
-def add_sponsorship_info_to_context(context: Dict[str, Any], user_profile: UserProfile) -> None:
-    def key_helper(d: Any) -> int:
-        return d[1]["display_order"]
-
-    context.update(
-        realm_org_type=user_profile.realm.org_type,
-        sorted_org_types=sorted(
-            (
-                [org_type_name, org_type]
-                for (org_type_name, org_type) in Realm.ORG_TYPES.items()
-                if not org_type.get("hidden")
-            ),
-            key=key_helper,
-        ),
-    )
 
 
 @zulip_login_required
@@ -47,30 +21,12 @@ def add_sponsorship_info_to_context(context: Dict[str, Any], user_profile: UserP
 def sponsorship_request(request: HttpRequest) -> HttpResponse:
     user = request.user
     assert user.is_authenticated
-    context: Dict[str, Any] = {}
 
-    customer = get_customer_by_realm(user.realm)
-    if customer is not None and customer.sponsorship_pending:
-        if is_realm_on_paid_plan(user.realm):
-            return HttpResponseRedirect(reverse("billing_home"))
+    billing_session = RealmBillingSession(user)
+    context = billing_session.get_sponsorship_request_context()
+    if context is None:
+        return HttpResponseRedirect(reverse("billing_home"))
 
-        context["is_sponsorship_pending"] = True
-
-    if user.realm.plan_type == user.realm.PLAN_TYPE_STANDARD_FREE:
-        context["is_sponsored"] = True
-
-    if customer is not None:
-        plan = get_current_plan_by_customer(customer)
-        if plan is not None:
-            context["plan_name"] = plan.name
-            context["free_trial"] = plan.is_free_trial()
-        # We don't create CustomerPlan objects for fully sponsored realms via support page.
-        elif user.realm.plan_type == Realm.PLAN_TYPE_STANDARD_FREE:
-            context["plan_name"] = "Zulip Cloud Standard"
-        else:
-            context["plan_name"] = "Zulip Cloud Free"
-
-    add_sponsorship_info_to_context(context, user)
     return render(request, "corporate/sponsorship.html", context=context)
